requirements.txt
awscli>=1.18.216
boto3>=1.16.56
botocore>=1.19.56
sagemaker>=2.29.0
smdebug>=1.0.1
sagemaker-experiments>=0.1.26
PyAthena>=2.1.0
awswrangler

TA-Lib
holidays


python-dotenv
torch

# pygame latest version
pygame>=2.1.3


alpaca_trade_api>=2.1.0
ccxt>=1.66.32  # make sure that >=1.66.32
elegantrl
exchange_calendars==3.6.3 # because raise exception with 4.1.1, success tested with 3.6.3
gputil
gym==0.26.2
gymnasium==0.29.1
importlib-metadata==4.13.0
jqdatasdk
lz4
# plot
matplotlib
# data handling
numpy>=1.17.3
pandas>=1.1.5

#hooks
pre-commit
pyfolio

# testing requirements
pytest
ray[default]
ray[tune]
recommonmark
recommonmark

# Model Building Requirements
scikit-learn>=0.21.0

# packaging
#setuptools>=41.4.0
setuptools>=65.5.0


# to build docs using sphinx
sphinx


# to build docs using sphinx
sphinx
sphinx_rtd_theme
sphinx_rtd_theme


SQLAlchemy
stable-baselines3[extra]
stockstats>=0.4.0
swig


tensorboardX
wheel>=0.33.6
wrds

# market data & paper trading API
yfinance

# deployment
uvicorn
fastapi